+++ db/migrations/sharding_3/20251230045649_initial_schema.sql
@@ -1,7 +1,7 @@
 -- Create "dm_posts_016" table
 CREATE TABLE `dm_posts_016` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -14,8 +14,8 @@ CREATE INDEX `idx_dm_posts_016_user_id` ON `dm_posts_016` (`user_id`);
 CREATE INDEX `idx_dm_posts_016_created_at` ON `dm_posts_016` (`created_at`);
 -- Create "dm_posts_017" table
 CREATE TABLE `dm_posts_017` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -28,8 +28,8 @@ CREATE INDEX `idx_dm_posts_017_user_id` ON `dm_posts_017` (`user_id`);
 CREATE INDEX `idx_dm_posts_017_created_at` ON `dm_posts_017` (`created_at`);
 -- Create "dm_posts_018" table
 CREATE TABLE `dm_posts_018` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -42,8 +42,8 @@ CREATE INDEX `idx_dm_posts_018_user_id` ON `dm_posts_018` (`user_id`);
 CREATE INDEX `idx_dm_posts_018_created_at` ON `dm_posts_018` (`created_at`);
 -- Create "dm_posts_019" table
 CREATE TABLE `dm_posts_019` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -56,8 +56,8 @@ CREATE INDEX `idx_dm_posts_019_user_id` ON `dm_posts_019` (`user_id`);
 CREATE INDEX `idx_dm_posts_019_created_at` ON `dm_posts_019` (`created_at`);
 -- Create "dm_posts_020" table
 CREATE TABLE `dm_posts_020` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -70,8 +70,8 @@ CREATE INDEX `idx_dm_posts_020_user_id` ON `dm_posts_020` (`user_id`);
 CREATE INDEX `idx_dm_posts_020_created_at` ON `dm_posts_020` (`created_at`);
 -- Create "dm_posts_021" table
 CREATE TABLE `dm_posts_021` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -84,8 +84,8 @@ CREATE INDEX `idx_dm_posts_021_user_id` ON `dm_posts_021` (`user_id`);
 CREATE INDEX `idx_dm_posts_021_created_at` ON `dm_posts_021` (`created_at`);
 -- Create "dm_posts_022" table
 CREATE TABLE `dm_posts_022` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -98,8 +98,8 @@ CREATE INDEX `idx_dm_posts_022_user_id` ON `dm_posts_022` (`user_id`);
 CREATE INDEX `idx_dm_posts_022_created_at` ON `dm_posts_022` (`created_at`);
 -- Create "dm_posts_023" table
 CREATE TABLE `dm_posts_023` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -112,7 +112,7 @@ CREATE INDEX `idx_dm_posts_023_user_id` ON `dm_posts_023` (`user_id`);
 CREATE INDEX `idx_dm_posts_023_created_at` ON `dm_posts_023` (`created_at`);
 -- Create "dm_users_016" table
 CREATE TABLE `dm_users_016` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -123,7 +123,7 @@ CREATE TABLE `dm_users_016` (
 CREATE UNIQUE INDEX `idx_dm_users_016_email` ON `dm_users_016` (`email`);
 -- Create "dm_users_017" table
 CREATE TABLE `dm_users_017` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -134,7 +134,7 @@ CREATE TABLE `dm_users_017` (
 CREATE UNIQUE INDEX `idx_dm_users_017_email` ON `dm_users_017` (`email`);
 -- Create "dm_users_018" table
 CREATE TABLE `dm_users_018` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -145,7 +145,7 @@ CREATE TABLE `dm_users_018` (
 CREATE UNIQUE INDEX `idx_dm_users_018_email` ON `dm_users_018` (`email`);
 -- Create "dm_users_019" table
 CREATE TABLE `dm_users_019` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -156,7 +156,7 @@ CREATE TABLE `dm_users_019` (
 CREATE UNIQUE INDEX `idx_dm_users_019_email` ON `dm_users_019` (`email`);
 -- Create "dm_users_020" table
 CREATE TABLE `dm_users_020` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -167,7 +167,7 @@ CREATE TABLE `dm_users_020` (
 CREATE UNIQUE INDEX `idx_dm_users_020_email` ON `dm_users_020` (`email`);
 -- Create "dm_users_021" table
 CREATE TABLE `dm_users_021` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -178,7 +178,7 @@ CREATE TABLE `dm_users_021` (
 CREATE UNIQUE INDEX `idx_dm_users_021_email` ON `dm_users_021` (`email`);
 -- Create "dm_users_022" table
 CREATE TABLE `dm_users_022` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -189,7 +189,7 @@ CREATE TABLE `dm_users_022` (
 CREATE UNIQUE INDEX `idx_dm_users_022_email` ON `dm_users_022` (`email`);
 -- Create "dm_users_023" table
 CREATE TABLE `dm_users_023` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
