+++ db/migrations/sharding_4/20251230045650_initial_schema.sql
@@ -1,7 +1,7 @@
 -- Create "dm_posts_024" table
 CREATE TABLE `dm_posts_024` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -14,8 +14,8 @@ CREATE INDEX `idx_dm_posts_024_user_id` ON `dm_posts_024` (`user_id`);
 CREATE INDEX `idx_dm_posts_024_created_at` ON `dm_posts_024` (`created_at`);
 -- Create "dm_posts_025" table
 CREATE TABLE `dm_posts_025` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -28,8 +28,8 @@ CREATE INDEX `idx_dm_posts_025_user_id` ON `dm_posts_025` (`user_id`);
 CREATE INDEX `idx_dm_posts_025_created_at` ON `dm_posts_025` (`created_at`);
 -- Create "dm_posts_026" table
 CREATE TABLE `dm_posts_026` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -42,8 +42,8 @@ CREATE INDEX `idx_dm_posts_026_user_id` ON `dm_posts_026` (`user_id`);
 CREATE INDEX `idx_dm_posts_026_created_at` ON `dm_posts_026` (`created_at`);
 -- Create "dm_posts_027" table
 CREATE TABLE `dm_posts_027` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -56,8 +56,8 @@ CREATE INDEX `idx_dm_posts_027_user_id` ON `dm_posts_027` (`user_id`);
 CREATE INDEX `idx_dm_posts_027_created_at` ON `dm_posts_027` (`created_at`);
 -- Create "dm_posts_028" table
 CREATE TABLE `dm_posts_028` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -70,8 +70,8 @@ CREATE INDEX `idx_dm_posts_028_user_id` ON `dm_posts_028` (`user_id`);
 CREATE INDEX `idx_dm_posts_028_created_at` ON `dm_posts_028` (`created_at`);
 -- Create "dm_posts_029" table
 CREATE TABLE `dm_posts_029` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -84,8 +84,8 @@ CREATE INDEX `idx_dm_posts_029_user_id` ON `dm_posts_029` (`user_id`);
 CREATE INDEX `idx_dm_posts_029_created_at` ON `dm_posts_029` (`created_at`);
 -- Create "dm_posts_030" table
 CREATE TABLE `dm_posts_030` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -98,8 +98,8 @@ CREATE INDEX `idx_dm_posts_030_user_id` ON `dm_posts_030` (`user_id`);
 CREATE INDEX `idx_dm_posts_030_created_at` ON `dm_posts_030` (`created_at`);
 -- Create "dm_posts_031" table
 CREATE TABLE `dm_posts_031` (
-  `id` integer NOT NULL,
-  `user_id` integer NOT NULL,
+  `id` bigint NOT NULL,
+  `user_id` bigint NOT NULL,
   `title` text NOT NULL,
   `content` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -112,7 +112,7 @@ CREATE INDEX `idx_dm_posts_031_user_id` ON `dm_posts_031` (`user_id`);
 CREATE INDEX `idx_dm_posts_031_created_at` ON `dm_posts_031` (`created_at`);
 -- Create "dm_users_024" table
 CREATE TABLE `dm_users_024` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -123,7 +123,7 @@ CREATE TABLE `dm_users_024` (
 CREATE UNIQUE INDEX `idx_dm_users_024_email` ON `dm_users_024` (`email`);
 -- Create "dm_users_025" table
 CREATE TABLE `dm_users_025` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -134,7 +134,7 @@ CREATE TABLE `dm_users_025` (
 CREATE UNIQUE INDEX `idx_dm_users_025_email` ON `dm_users_025` (`email`);
 -- Create "dm_users_026" table
 CREATE TABLE `dm_users_026` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -145,7 +145,7 @@ CREATE TABLE `dm_users_026` (
 CREATE UNIQUE INDEX `idx_dm_users_026_email` ON `dm_users_026` (`email`);
 -- Create "dm_users_027" table
 CREATE TABLE `dm_users_027` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -156,7 +156,7 @@ CREATE TABLE `dm_users_027` (
 CREATE UNIQUE INDEX `idx_dm_users_027_email` ON `dm_users_027` (`email`);
 -- Create "dm_users_028" table
 CREATE TABLE `dm_users_028` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -167,7 +167,7 @@ CREATE TABLE `dm_users_028` (
 CREATE UNIQUE INDEX `idx_dm_users_028_email` ON `dm_users_028` (`email`);
 -- Create "dm_users_029" table
 CREATE TABLE `dm_users_029` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -178,7 +178,7 @@ CREATE TABLE `dm_users_029` (
 CREATE UNIQUE INDEX `idx_dm_users_029_email` ON `dm_users_029` (`email`);
 -- Create "dm_users_030" table
 CREATE TABLE `dm_users_030` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
@@ -189,7 +189,7 @@ CREATE TABLE `dm_users_030` (
 CREATE UNIQUE INDEX `idx_dm_users_030_email` ON `dm_users_030` (`email`);
 -- Create "dm_users_031" table
 CREATE TABLE `dm_users_031` (
-  `id` integer NOT NULL,
+  `id` bigint NOT NULL,
   `name` text NOT NULL,
   `email` text NOT NULL,
   `created_at` datetime NOT NULL,
